+++ field_friend/hardware/field_friend_hardware.py
@@ -5,7 +5,7 @@ import rosys
 
 import config.config_selection as config_selector
 
-from .axis_D1 import AxisD1
+from .axis_d1 import AxisD1
 from .can_open_master import CanOpenMasterHardware
 from .chain_axis import ChainAxisHardware
 from .double_wheels import DoubleWheelsHardware
